pyproject.toml
[project]
name = "proalgotrader-core-template"
version = "0.1.0"
description = "Algorithmic trading template with FastAPI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "httpx>=0.28.1",
    "websockets>=12.0",
    "python-dotenv>=1.1.0",
    "itsdangerous>=2.2.0",
    "pydantic>=2.10.6",
    "sqlmodel>=0.0.38",
    "pyotp>=2.9.0",
    "logzero>=1.7.0",
    "polars>=1.39.3",
    "pyarrow>=23.0.1",
    "pytz>=2026.1.post1",
    "python-fasthtml>=0.13.3",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "pyright>=1.1.397",
    "ruff>=0.11.0",
    "pre-commit>=4.5.0",
]

[tool.black]
line-length = 88
target-version = ["py313"]
include = '\.pyi?$'
exclude = '''
    /(
        \.git
        | \.venv
        | \.eggs
        | \.tox
        | build
        | dist
        | __pycache__
    )/
'''

[tool.ruff]
line-length = 88
target-version = "py313"
lint.select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
lint.ignore = ["E501", "B008", "B904", "B007", "E722", "N811", "UP042"]
exclude = [".git", ".venv", "__pycache__", "build", "dist"]

[tool.ruff.lint.per-file-ignores]
"app/fasthtml_views/**/*.py" = ["F403", "F405"]

[tool.pyright]
include = ["app", "proalgotrader_core", "db"]
exclude = [".venv", "**/node_modules", "build", "dist"]
pythonVersion = "3.13"
typeCheckingMode = "off"
useLibraryCodeForTypes = true
reportMissingImports = true
reportMissingTypeStubs = false

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
